+++ src/parser/Parser.cpp
@@ -6,11 +6,12 @@
 /*   By: Dugonzal <dugonzal@student.42urduliz.com>  +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/03/03 13:36:19 by Dugonzal          #+#    #+#             */
-/*   Updated: 2024/04/01 12:59:14 by Dugonzal         ###   ########.fr       */
+/*   Updated: 2024/04/01 18:06:51 by Dugonzal         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 # include "../../inc/parser/Parser.hpp"
+#include <cstddef>
 
 Parser::Parser(void) { }
 
@@ -18,11 +19,12 @@ Parser::~Parser(void) { delete [] dataServers; }
 
 Parser::Parser(const string &filename): BaseParser(filename) {
   dataServers = new vector<string>[nServers];
-  unsigned int i = -1;
+  size_t i = -1;
+
   for (vector<string>::const_iterator it = data.begin(); \
     it != data.end() && ++i < nServers; it++) {
       if (it->find("server") != string::npos) {
-        while (++it != data.end() && it->find("};") == string::npos) 
+        while (++it != data.end() && it->find("};") == string::npos)
           dataServers[i].push_back(*it);
       }
   }
